+++ src/background/services/zkProof/protocols/IZkProof.ts
@@ -1,0 +1,5 @@
+import ZkIdentityDecorater from "@src/background/identityDecorater";
+
+export interface IZkProof<Request, Return> {
+    genProof(identity: ZkIdentityDecorater, request: Request): Promise<Return>;
+}
