+++ src/background/services/zkProof/protocols/utils.ts
